ServoingTask: Do not create fake obstacles in front of the robot
For security reasons, we mark unknown areas near the robot as obstacle.
This sometimes created fake obstacles in front of the robot on inclinations.
To get rid of this, we now mark unknown areas beside and behind the robot
as obstacle, but not in front any more. We still mark everything "under"
the robot as obstacle.