Reworked trajectory following frame logic
Until this commit, trajectory following was only possible,
if we constantly got updates for the body2GlobalTrajectory
transformation. With this commit this behaviour is changed.
Every time a body2GlobalTrajectory is received, the transformation
from globalTrajectory to map is computed. The following of
the trajectory is then done in the map frame.

This is especially useful, if your global trajectory frame
has a slow update (or is static). If the global trajectory frame
has a high update rate, this should almost make no difference.